+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # @grogarden/util
 
+## 0.18.3
+
+### Patch Changes
+
+- return headers from `fetch_value` ([#24](https://github.com/grogarden/util/pull/24))
+
+## 0.18.2
+
+### Patch Changes
+
+- add a default 'content-type' header for `fetch_value` ([2126d03](https://github.com/grogarden/util/commit/2126d03))
+
 ## 0.18.1
 
 ### Patch Changes